+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 1.2.5 (2019-05-17)
+ * Added some options types
  * Added get/set for options
  
 ## 1.2.4 (2019-05-16)